backend/requirements.txt
asgiref==3.5.0
attr==0.3.1
backports.zoneinfo==0.2.1
black==22.3.0
click==8.1.2
cycler==0.11.0
Django==4.0.3
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
fonttools==4.30.0
gunicorn==20.1.0
html5lib==1.1
joblib==1.1.0
kiwisolver==1.4.0
matplotlib==3.5.1
mypy-extensions==0.4.3
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
psycopg2-binary==2.9.3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tk==0.1.0
tomli==2.0.1
typing_extensions==4.1.1
webencodings==0.5.1

